+++ async-race/src/view/view.ts
